requirements.txt
# Existing dependencies from pyproject.toml
torch>=1.12.1
opencv-python>=4.6.0.66
numpy>=1.23.4
Pillow>=9.2.0
matplotlib>=3.6.1
pandas>=1.5.0
torchvision>=0.13.1
PyYAML>=6.0
tqdm>=4.64.1
seaborn>=0.12.0
ipython>=8.5.0

# Additional dependencies for enhanced tactical analysis
scipy>=1.9.2
# For convex hull calculations in formation polygons

# Optional: For even better tactical analysis
scikit-learn>=1.1.0
# For clustering analysis of player positions (future enhancement)